Update patch set 3
Patch Set 3:

It does the job (exposes the preferred network modes for each RIL instance), but I would prefer not to have a custom RIL just for this and instead (if that's possible) changing the way audio routing is handled for HTC directly on android_hardware_qcom_audio.

I don't understand why these props are not available in cm-13.0, even if I'm still using L RIL blobs.

Patch-set: 3